根据ai建议重命名部分方法

+++ sonsure-dumper-test/src/test/java/com/sonsure/dumper/test/jdbc/OracleJdbcTemplateDaoImplTest.java
@@ -35,7 +35,7 @@ public class OracleJdbcTemplateDaoImplTest {
         OracleUser oracleUser = new OracleUser();
         oracleUser.setUsername("liyd");
 
-        Long id = (Long) jdbcDao.executeInsert(oracleUser);
+        Long id = jdbcDao.executeInsert(oracleUser);
 
         Assertions.assertNotNull(id);
     }
